% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlotSBCovidCases.R
\name{simplePlotSBCovidCases}
\alias{simplePlotSBCovidCases}
\title{Daily Santa Barbara Covid data from the NY Times
simplePlotSBCovidCases makes a plot}
\usage{
simplePlotSBCovidCases()
}
\description{
Daily Santa Barbara Covid data from the NY Times
simplePlotSBCovidCases makes a plot
}
\examples{
simplePlotSBCases
}
